requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
PyYAML>=6.0.1
openai>=1.51.0
python-dotenv>=1.0.1
pytest>=8.3.0
